requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
google-generativeai>=0.8.0
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.23
